:sparkles: feat: Open a problem and terminal in a new tab
Opening a problem source code is a frequently accomplished task and
everytime a terminal window needs to be opended too alongside it in
order compile/run the problem. This new feature let's the user specify a
problem source file path in a command and that file will be opened in a
new tab along with a new vertically splitted terminal window.

Configurability is kept to minimal for now:

 * Users cannot change the orientation of the terminal to horizontal.

 * Also the command cannot yet intelligently deduce the file path from
   contest/problem codes, the file path has to specified explicitly.

These features might be added in some future commit.